requirements.txt
langchain
langchain-core
langchain-gemini
langchain-openai
langchain-text-splitters
langchain-community
python-dotenv
langchain-google-genai
streamlit
pypdf
langchain-chroma